🎚️ Enabled server config & modified level system to suit it

+++ src/commands/Utilities/ping.js
@@ -5,6 +5,7 @@ export default {
 	alias: [''],
 
 	async run(msg, args, author, isInteraction = false) {
+		if (isInteraction) return {content: 'Oopsie, send `g!ping` in chat to check my ping..'}
 		const newMsg = await msg.reply({content:'🧮 Calculating ping...'})
 		const botLatency = newMsg.createdTimestamp - msg.createdTimestamp
 		await newMsg.delete();
